requirements.txt
fastapi 
"uvicorn[standard]" 
pydantic 
requests 
langchain 
langchain-openai 
langchain-core 
langchain-community 
faiss-cpu 
openai 
python-dotenv
sqlalchemy
asyncpg
psycopg2-binary
